moved previous homepage components to customers folder
